python-function/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform
opencensus-extension-azure-functions
opencensus-ext-requests
azure-functions
